Add files via upload
Assignment 1

Q1. The volume of a sphere with radius r is (4/3)*π*r^3. What is the volume of a sphere with radius 5?
Q2. We have v=u+at where v is velocity, u is initial velocity, and t is the time. And value for each is as follows: v=25m/s. u=0m/s, t=10 seconds. Find the acceleration a. [Rearrange the equation, also use proper formatting syntax while displaying the answer.]

Q3. Take an arbitrary input string from user and print it 10 times.
Q4. Write a python script to demonstrate at least 10 built in functions in Python.
Q5. Write a Python program to sum all the items in a list. Do not use built-in sum() function.
Q6. Write a Python program to replace the last element in a list with another list.

Sample data : [1, 3, 5, 7, 9, 10], [2, 4, 6, 8]
Expected Output: [1, 3, 5, 7, 9, 2, 4, 6, 8]
Q7. Write a Python program to calculate a dog's age in dog's years. 
Note: For the first two years, a dog year is equal to 10.5 human years. After that, each  dog year equals 4 human years. 
Expected Output: 
Input a dog's age in human years: 15  
The dog's age in dog's years is 73 

Q8Write a Python program to add an item in a tuple.
Q9. Write a Python program to calculate the length of a string. Do not use built-in len() function.
Q10. Write a Python program to print a specified list after removing the 0th, 4th and 5th elements. Sample List : ['Red', 'Green', 'White', 'Black', 'Pink', 'Yellow'] 
Expected Output : ['Green', 'White', 'Black']